feat!: Update fsatctl communication protocol with services
This commit updates the communication protocol with services. It now
uses json as a serialization protocol, chosen for more flexibility, it
also changes the socket type to stream, for better communication
capabilities with the services.
Another update was in the command line parsing, which was changed to be
getopt based instead of the messy way from before.